feat: 댓글 리팩토링 및 수정 삭제 권한 설정

+++ src/main/java/com/YYduo/KkuldongVarietyStore/domain/comment/service/CommentService.java
@@ -6,12 +6,14 @@ import com.YYduo.KkuldongVarietyStore.domain.comment.entity.Comment;
 import com.YYduo.KkuldongVarietyStore.domain.comment.mapper.CommentMapper;
 import com.YYduo.KkuldongVarietyStore.domain.comment.repository.CommentRepository;
 import com.YYduo.KkuldongVarietyStore.domain.diary.entity.Diary;
+import com.YYduo.KkuldongVarietyStore.domain.diary.mapper.DiaryMapper;
 import com.YYduo.KkuldongVarietyStore.domain.diary.repository.DiaryRepository;
 import com.YYduo.KkuldongVarietyStore.domain.member.entity.Member;
 import com.YYduo.KkuldongVarietyStore.domain.member.repository.MemberRepository;
 import com.YYduo.KkuldongVarietyStore.exception.CustomException;
 import com.YYduo.KkuldongVarietyStore.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 
 import javax.transaction.Transactional;
@@ -27,14 +29,21 @@ public class CommentService {
     private final MemberRepository memberRepository;
     private final CommentMapper commentMapper;
 
-    public Comment saveComment(CommentPostDto commentPostDto) {
-        Comment comment = commentMapper.commentPostDtoToComment(commentPostDto);
+    public Comment  saveComment(Comment comment) {
 
-        Diary diary = diaryRepository.findById(commentPostDto.getDiaryId())
+
+        Member member = memberRepository.findById(comment.getMember().getId())
+                .orElseThrow(() -> new CustomException(ExceptionCode.MEMBER_NOT_FOUND));
+
+        Diary diary = diaryRepository.findById(comment.getDiary().getId())
                 .orElseThrow(() -> new CustomException(ExceptionCode.DIARY_NOT_FOUND));
 
-        Member member = memberRepository.findById(commentPostDto.getMemberId())
-                .orElseThrow(() -> new CustomException(ExceptionCode.MEMBER_NOT_FOUND));
+
+
+
+//        Comment comment = commentMapper.INSTANCE.commentPostDtoToComment(commentPostDto);
+
+        comment.setAuthor(member.getNickname());
 
         comment.setDiary(diary);
         comment.setMember(member);
@@ -44,18 +53,27 @@ public class CommentService {
 
 
 
-    public Comment updateComment(Long commentId, CommentPatchDto commentPatchDto) {
+    public Comment updateComment(Long commentId, CommentPatchDto commentPatchDto, Member auth) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.COMMENT_NOT_FOUND));
+
+        // Check if the authenticated user is the author of the comment
+        if (!comment.getMember().getId().equals(auth.getId())) {
+            throw new CustomException(ExceptionCode.UNAUTHORIZED_ACCESS);
+        }
 
         commentMapper.commentPatchDtoToComment(comment, commentPatchDto);
 
         return commentRepository.save(comment);
     }
 
-    public void deleteComment(Long commentId) {
+
+    public void deleteComment(Long commentId, Member auth) {
         Comment comment = commentRepository.findById(commentId)
                 .orElseThrow(() -> new CustomException(ExceptionCode.COMMENT_NOT_FOUND));
+        if (!comment.getMember().getId().equals(auth.getId())) {
+            throw new CustomException(ExceptionCode.UNAUTHORIZED_ACCESS);
+        }
         commentRepository.delete(comment);
     }
